Prog:Lab1: change measurement of x and y

+++ prog/01_oop_classes/src/point.cpp
@@ -1,14 +1,14 @@
 #include "../include/point.h"
 
-Point::Point(int x_val, int y_val): x(x_val), y(y_val) {}
+Point::Point(double x_val, double y_val): x(x_val), y(y_val) {}
 
 Point::Point(const Point &other): x(other.x), y(other.y) {}
 
 Point::~Point() {}
 
-int Point::x_val() const {return x;}
+double Point::x_val() const {return x;}
 
-int Point::y_val() const {return y;}
+double Point::y_val() const {return y;}
 
 Point& Point::operator=(const Point &other) {
     if (&other == this) {
